refactor(front): move getWhitelistedProviders to lib/api/assistant/models
