+++ front/lib/api/assistant/conversation/title.ts
@@ -1,11 +1,9 @@
 import type { AgentActionSpecification } from "@app/lib/actions/types/agent";
 import { runMultiActionsAgent } from "@app/lib/api/assistant/call_llm";
 import { renderConversationAsText } from "@app/lib/api/assistant/conversation/render_as_text";
+import { getWhitelistedProviders } from "@app/lib/api/assistant/models";
 import { publishConversationEvent } from "@app/lib/api/assistant/streaming/events";
-import {
-  getSmallWhitelistedModel,
-  getWhitelistedProviders,
-} from "@app/lib/assistant";
+import { getSmallWhitelistedModel } from "@app/lib/assistant";
 import type { AuthenticatorType } from "@app/lib/auth";
 import { Authenticator } from "@app/lib/auth";
 import { ConversationResource } from "@app/lib/resources/conversation_resource";
